+++ modules/common/src/memcpy_s.c
@@ -13,7 +13,7 @@
 
 #include "oscore_edhoc_error.h"
 
-enum err check_buffer_size(uint64_t is_size, uint64_t required_size)
+enum err check_buffer_size(uint32_t is_size, uint32_t required_size)
 {
 	if (is_size < required_size) {
 		return buffer_to_small;
@@ -22,8 +22,8 @@ enum err check_buffer_size(uint64_t is_size, uint64_t required_size)
 	}
 }
 
-enum err _memcpy_s(uint8_t *dest, uint64_t destSize, const uint8_t *src,
-		   uint64_t count)
+enum err _memcpy_s(uint8_t *dest, uint32_t destSize, const uint8_t *src,
+		   uint32_t count)
 {
 	TRY(check_buffer_size(destSize, count));
 	memcpy(dest, src, count);
