+++ modules/edhoc/inc/ciphertext.h
@@ -14,18 +14,20 @@
 enum ciphertext { CIPHERTEXT2, CIPHERTEXT3, CIPHERTEXT4 };
 
 enum err ciphertext_gen(enum ciphertext ctxt, struct suite *suite,
-				uint8_t *id_cred, uint32_t id_cred_len,
-				uint8_t *signature_or_mac,
-				uint32_t signature_or_mac_len, uint8_t *ead,
-				uint32_t ead_len, uint8_t *prk, uint8_t prk_len,
-				uint8_t *th, uint8_t th_len,
-				uint8_t *ciphertext, uint32_t *ciphertext_len);
+			uint8_t *id_cred, uint32_t id_cred_len,
+			uint8_t *signature_or_mac,
+			uint32_t signature_or_mac_len, uint8_t *ead,
+			uint32_t ead_len, uint8_t *prk, uint32_t prk_len,
+			uint8_t *th, uint32_t th_len, uint8_t *ciphertext,
+			uint32_t *ciphertext_len);
 
-enum err ciphertext_decrypt_split(
-	enum ciphertext ctxt, struct suite *suite, uint8_t *prk,
-	uint8_t prk_len, uint8_t *th, uint8_t th_len, uint8_t *ciphertext,
-	uint32_t ciphertext_len, uint8_t *id_cred, uint32_t *id_cred_len,
-	uint8_t *signature_or_mac, uint32_t *signature_or_mac_len, uint8_t *ead,
-	uint32_t *ead_len);
+enum err ciphertext_decrypt_split(enum ciphertext ctxt, struct suite *suite,
+				  uint8_t *prk, uint8_t prk_len, uint8_t *th,
+				  uint8_t th_len, uint8_t *ciphertext,
+				  uint32_t ciphertext_len, uint8_t *id_cred,
+				  uint32_t *id_cred_len,
+				  uint8_t *signature_or_mac,
+				  uint32_t *signature_or_mac_len, uint8_t *ead,
+				  uint32_t *ead_len);
 
 #endif
